feat : new endpoint find customer credit

+++ internal/domain/customer/customer.go
@@ -7,12 +7,12 @@ import (
 )
 
 type Customer struct {
-	ID            string  `json:"id"`
-	Code          string  `json:"code"`
-	Name          string  `json:"name"`
-	Description   string  `json:"description"`
-	Active        bool    `json:"active"`
-	InitialCredit float64 `json:"initialCredit"`
+	ID            string `json:"id"`
+	Code          string `json:"code"`
+	Name          string `json:"name"`
+	Description   string `json:"description"`
+	Active        bool   `json:"active"`
+	InitialCredit int64  `json:"initialCredit"`
 }
 
 type AddPriceRequest struct {
